:alembic: Make animations independent from the Dragonfly tick rate

+++ src/main/java/net/inceptioncloud/dragonfly/engine/sequence/types/DoubleSequence.kt
@@ -2,14 +2,14 @@ package net.inceptioncloud.dragonfly.engine.sequence.types
 
 import net.inceptioncloud.dragonfly.engine.sequence.Sequence
 
-class DoubleSequence(from: Double, to: Double, duration: Int) : Sequence<Double>(from, to, duration) {
+class DoubleSequence(from: Double, to: Double, duration: Long) : Sequence<Double>(from, to, duration) {
     /**
      * A function to interpolate the value.
      *
      * Responsible for providing the transition between the [from] and the [to] value by
      * interpolating it.
      *
-     * @param progress the the quotient of the [time] and the [duration] transformed by the [easing] function
+     * @param progress the the quotient of the [expiredTime] and the [duration] transformed by the [easing] function
      */
     override fun interpolate(progress: Double): Double {
         return (to - from) * progress + from
